+++ passes/S019/testdata/src/a/schema/schema.go
@@ -1,0 +1,7 @@
+package schema
+
+type Schema struct {
+	Computed bool
+	Optional bool
+	Required bool
+}