+++ CHANGELOG.md
@@ -14,6 +14,7 @@ FEATURES
 * **New Check:** `S016`: check for `Schema` that `Set` is only configured for `TypeSet`
 * **New Check:** `S017`: check for `Schema` that `MaxItems` and `MinItems` are only configured for `TypeList`, `TypeMap`, or `TypeSet`
 * **New Check:** `S018`: check for `Schema` that should prefer `TypeList` with `MaxItems: 1`
+* **New Check:** `S019`: check for `Schema` that should omit `Computed`, `Optional`, or `Required` set to `false`
 
 BUG FIXES